Replay: added --param-file option
