+++ Tools/Replay/LogReader.h
@@ -1,11 +1,12 @@
 #include "VehicleType.h"
 #include "DataFlashFileReader.h"
 #include "LR_MsgHandler.h"
+#include "Parameters.h"
 
 class LogReader : public DataFlashFileReader
 {
 public:
-    LogReader(AP_AHRS &_ahrs, AP_InertialSensor &_ins, AP_Baro &_baro, Compass &_compass, AP_GPS &_gps, AP_Airspeed &_airspeed, DataFlash_Class &_dataflash, const struct LogStructure *structure, uint8_t num_types, const char **&nottypes);
+    LogReader(AP_AHRS &_ahrs, AP_InertialSensor &_ins, AP_Baro &_baro, Compass &_compass, AP_GPS &_gps, AP_Airspeed &_airspeed, DataFlash_Class &_dataflash, const char **&nottypes);
     bool wait_type(const char *type);
 
     const Vector3f &get_attitude(void) const { return attitude; }
@@ -62,9 +63,6 @@ private:
 
     // next available msgid for mapping
     uint8_t next_msgid = 1;
-
-    const struct LogStructure *structure;
-    uint8_t num_types;
 
     LR_MsgHandler::CheckState check_state;
